LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/body-onerror-runtime-error-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable

PASS <body onerror> - runtime error in <script>
PASS <body onerror> - runtime error in <script> (column)
